refactor: move builders into follow_ref module, and reorganise sub modules
