+++ src/follow_refs/build.rs
@@ -1,0 +1,16 @@
+use winnow::{BStr, PResult};
+
+use super::Builder;
+
+/// Generalisation on the [`Extract`] trait, which allows the extraction logic to follow references.
+pub trait Build<'de>: Sized {
+    /// Build an object that rely on a reference, which would be instantiated with the help of the
+    /// supplied `builder`.
+    ///
+    /// The [`Build`] trait, like the [`Extract`] trait, is a linear parser above all, hence we
+    /// supply an `input`. References found during parsing, if any, are first parsed as such, and
+    /// then instantiated by the `builder`.
+    fn build<B>(input: &mut &'de BStr, builder: &B) -> PResult<Self>
+    where
+        B: Builder<'de>;
+}
