+++ src/follow_refs/mod.rs
@@ -41,9 +41,8 @@
 //! a PDF document, regardless of whether some fields may be represented as references in the
 //! serialisation.
 
-mod behaviour;
-mod builders;
-mod parser;
+mod build;
+mod builder;
 
-pub use behaviour::{Build, Builder};
-pub use parser::{BuilderParser, LivreBuilder};
+pub use build::Build;
+pub use builder::{Builder, BuilderParser, LivreBuilder};
